+++ cmplibrary/src/main/java/com/sourcepoint/gdpr_cmplibrary/ConsentAction.java
@@ -2,6 +2,9 @@ package com.sourcepoint.gdpr_cmplibrary;
 
 import org.json.JSONObject;
 
+import java.util.HashMap;
+import java.util.Map;
+
 public class ConsentAction {
 
 
@@ -10,6 +13,7 @@ public class ConsentAction {
     public final String privacyManagerId;
     public final boolean requestFromPm;
     public final JSONObject pmSaveAndExitVariables;
+    private Map pubData = new HashMap();
 
     ConsentAction(int actionType, String choiceId, String privacyManagerId, boolean requestFromPm, JSONObject pmSaveAndExitVariables) {
         this.actionType = ActionTypes.valueOf(actionType);
@@ -23,9 +27,12 @@ public class ConsentAction {
         this(actionType, choiceId, null, requestFromPm, pmSaveAndExitVariables);
     }
 
-    public static ConsentAction getEmptyDismissAction(boolean isPmOn) {
-        return isPmOn ? new ConsentAction(ActionTypes.PM_DISMISS.code, "", "", true, new JSONObject())
-                : new ConsentAction(ActionTypes.MSG_CANCEL.code, "", "", false, new JSONObject());
+    public void setPubData(Map pubData){
+        this.pubData = pubData;
+    }
+
+    public JSONObject getPubData(){
+        return new JSONObject(pubData);
     }
 
 }