+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 5.3.8 (September, 9, 2020)
+* add `setOnNoIntentActivitiesFound(String intentUrl)` to avoid crashing when no url intent handler is found.
+* prepare for the new  `pubData` feature. Ability to set `Set<Object> pubData` on the `ConsentAction` obj before sending consent:
+```JAVA
+    .setOnBeforeSendingConsent((consentAction, consentHandler) -> {
+        consentAction.setPubData(Collections.singletonMap("foo", "bar"));
+        consentHandler.post(consentAction);
+    })
+```
+
+
 ## 5.3.8 (August, 26, 2020)
 * fix parsing error on native msg causing app crash #249
 * fix onError called twice on some cases #246